Update patch set 2
Patch Set 2:

> yes, this was just a test patch not intended to be committed. That
 > being said we probably should add some infra so that we can change
 > config more easily that having top write these she scripts and also
 > a way to pass the values via the job.

it would be cool if we could tickle these configs with the REST API.

these scripts are not super painful, but they do add yet another
layer of complicated steps.

btw, we do have an example of passing a config value to a script
like this from a jenkins parameter. our SG mode:

https://github.com/opendaylight/integration-test/blob/master/csit/scripts/set_sg_mode.sh

Patch-set: 2