Update patch set 2
Patch Set 2:

> @jamo for the var passing to scripts... what I was wondering is if
 > there is a way to not have to update every single job that uses the
 > template so that the var could be passed. I guess it isn't so bad
 > with the defaults, like what we added for the vpp job to get it
 > working. netvirt would only need these vars so we don't touch the
 > other project yamls.
 

yeah, with good defaults at the template level we don't need to worry
about these for jobs that don't care. We probably have a lot of examples
for that already.

 > restconf would be nice, but we need to add code in netvirt to
 > handle the updates. currently that is not there. with restconf, we
 > would still need to define the vars though right?

yeah, at some level the value we want to configure something with needs
to be known. I prefer doing it all at the robot level somehow. It's
easier and more visible (imho).

Patch-set: 2