Update patch set 4
Patch Set 4: Code-Review-1

(13 comments)

use "exclude" for test cases you want skipped. it can replace, or be in addition to the
not-implemented tag already there. Our upstream pybot is already configured to not
run cases with the exclude tag

the file numbering scheme is not required. the testplan files we use to drive which 
suites run for which jobs can be tailored in any order you like. It's not a problem to
keep, just fyi

large patches are slow going. smaller patches will get reviewed and merged faster.
However, this was not the typical large patch. This was mostly documentation of
future test cases to come. I *REALLY* like this approach. I had brought kind of
idea to the integration team in the past. I plan to share this patch with them to
show the idea in practice. The IOTDM team can make this test plan in robot documentation
first and then let the implementation happen as possible. Good work :)

NOTE: I did not dig deep in to the actual test case documentation details. I would
expect that IOTDM engineers can do a better job at that, if they haven't already.

Patch-set: 4
Label: Code-Review=-1