Add distchanges lookup for aggregated merges
Some distributions end up with a aggregated merge
of a number of gerrits in a patch stack. The data included in the git.properties
includes all the gerrit's commit messages. This gerrit adds a
search using the commit message of the patch at the top of the stack. [1]
is an example of one of these types of gerrits.

[1] https://git.opendaylight.org/gerrit/#/c/51304/

Change-Id: I26663b44e08b2f09c18dfff98efe560fbeaae4e4
Signed-off-by: Sam Hague <shague@redhat.com>