Update patch set 8
Patch Set 8:

> (1 comment)

I tried to fix the problem like this:
	
	Verify And Collect VM DHCP Addresses
    [Arguments]    @{VM_INSTANCES}
    [Documentation]    Check if vm received ip address, if not sleep 5s and try again.
    ...    eventually prints console log of each vm that didn't received ip.
    ${VM_IPS}    ${DHCP_IP}    Verify VMs Received DHCP Lease    @{VM_INSTANCES}
    ${status}    ${message}    Should Not Contain    ${VM_IPS}    None
    ${status}    ${message}    Should Not Contain    ${DHCP_IP}    None
    [Return]    ${VM_IPS}    ${DHCP_IP}

It did loop when it had None value, but the Verify And Collect VM DHCP Addresses func failed right after it finished the 60s of the loop and it didn't insert values into @{NET1_VM_IPS}, @{NET2_VM_IPS} variables.
I saw that if I want it to ignore error I can add the command:Run Keyword And Ignore Error. 
But I didn't succeed to do combine the 2 commands - ignore error and return back the ips from the Verify And Collect VM DHCP Addresses function.
There's an easy way to do that?

Patch-set: 8