Update patch set 3
Patch Set 3: Code-Review-2

hold tight on this one. Luis pointed out in the bug 5173 that possibly
what I'm seeing in netvirt (and trying to hide with this patch) could
be a real/new bug to be fixed. will investigate and come back to this
if needed.

Patch-set: 3
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2