Update patch set 24
Patch Set 24: Code-Review-2

> The CSIT passes (the failure here is because it's running on boron)
 > 
 > https://jenkins.opendaylight.org/sandbox/job/genius-csit-1node-genius-all-carbon/4/
 > 
 > Can we merge this?

You can add a line like this (exactly this if you want) at the
beginning of the new v6 test cases:

    Pass Execution If    "${ODL_STREAM}" == "boron"    IPV6 support is not available in boron or before


that will let things be skipped (marked as PASS) in boron, but it will
run in everything else. ODL_STREAM is passed on the pybot command
line for you already.

I put a -1 because I think this is important to do

Patch-set: 24
Label: Code-Review=-2