Update patch set 3
Patch Set 3:

This seems like it will work, and it lets us keep our
model dump in place. The 20s delay will be overridden
by the robot requests library own 1s timeout. So, we still
will be slower when model dumps are failing, but not 20s
per request.

here is a sandbox run that ran with it patch set 2:
https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-3node-openstack-newton-jamo-upstream-transparent-carbon/2/

that was 2h48m. which is much faster than the 4+ (sometimes 6h to 
abort) we are seeing.

the one problem with patch set 2 was that if/when the model dump
timed out after 1s, it marked the test teardown as a failure.
patch set 3 should address that, and sandbox run for this is here:

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-3node-openstack-newton-jamo-upstream-transparent-carbon/3/

this is in conflict with what Alon was working on, but I wanted to
find a way to keep model dumps there if possible. Alon's patch is
here:

https://git.opendaylight.org/gerrit/#/c/52642/

Patch-set: 3
Label: Verified=0