Update patch set 9
Patch Set 9:

Mor,

where are we with this. I finally caught back up to this in my
gerrit queue. As things are improved a lot right now, do we 
need to keep iterating on this patch?

I just gave it a rebase, since several patches have come in 
recently. seems ok.

Patch-set: 9