Update patch set 9
Patch Set 9:

> > Mor,
 > >
 > > where are we with this. I finally caught back up to this in my
 > > gerrit queue. As things are improved a lot right now, do we
 > > need to keep iterating on this patch?
 > >
 > > I just gave it a rebase, since several patches have come in
 > > recently. seems ok.
 > 
 > We were trying to think how to fix Verify And Collect VM DHCP
 > Addresses func because right now it doesn't iterate in case of None
 > vaule, I offered a suggestion for a fix:

I'm not sure now. I *think* it's still iterating when None is seen.
this looks like it's checking up to 3m for all VMs to get ips right?

https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-newton-nodl-v2-upstream-stateful-carbon/151/archives/log.html.gz#s1-s1-s1-t7-k3


 > > Mor,
 > >
 > > where are we with this. I finally caught back up to this in my
 > > gerrit queue. As things are improved a lot right now, do we
 > > need to keep iterating on this patch?
 > >
 > > I just gave it a rebase, since several patches have come in
 > > recently. seems ok.
 > 
 > Right now it doesn't iterate when there's a None value, I offered a
 > way to fix it:
 > change Verify And Collect VM DHCP Addresses not to return ips
 > and change the command ${VM_IP_NET1}    ${DHCP_IP1}    Wait Until
 > Keyword Succeeds    60s    5s    Verify And Collect VM DHCP
 > Addresses    @{VM_INSTANCES_NET1}
 > to ${status}    ${message}    Run Keyword And Ignore Error    Wait
 > Until Keyword Succeeds    60s    5s    Verify And Collect VM DHCP
 > Addresses    @{VM_INSTANCES_NET1}
 > then add the commands:
 > ${VM_IP_NET1}    ${DHCP_IP1}    Verify VMs Received DHCP Lease
 > Set Suite Variable    ${VM_IP_NET1}
 > Set Suite Variable    ${VM_IP_NET2}
 > Set Suite Variable    @{VM_INSTANCES_NET1}
 > Set Suite Variable    @{VM_INSTANCES_NET2}
 > and in the end if status is FAIL print nova console log and fail
 > the test
 > what do you think?

ok. I follow this now. I like it.

but, do you like the actual names of these keywords? can we refactor the names
to be more accurate? and/or just have a single keyword to do this "ip checking/collecting".

what if it's a single keyword like "Collect VM IP Addresses" that takes a list of VMs and
a flag that tells it to fail if None is found, otherwise just return None in the list of
ips it returns.

so, when the flag is set to fail on None, it will work in the WUKS, then after your
two WUKS in your code above, you call it again without that flag set to fail and it
will gladly return a list of IPs even if some are 'None'

thanks for patiently pulling me along here :)

Patch-set: 9
Label: Verified=0