Update patch set 4
Patch Set 4:

Hi All,

Found out while testing that, ODL_L3 and L2Gateway functionalities are not co-existing in Mitaka (but works fine in newton)

Reported to N Vivekanandan and yet to conclude about the changes to be delivered for csit-3-instance-type job so that this works.

Otherwise the testcases pass with the current OpenstackOperations changes.

I shall update here once the Mitaka issue is solved. 

Thanks,
Abbas

Patch-set: 4