Update patch set 8
Patch Set 8: Code-Review+1

Ursicio,

is this ready to merge on your end? I know you have jjb patches
waiting on this too. just give a +1 if ready. feel free to 
test in sandbox first and put results here, but not a hard 
requirement for a new suite.

Patch-set: 8
Label: Code-Review=+1