Fix rest body of create-node
When we load all features,it loads ovsdb and hwvtep both
and both the yang models has connection-info container in
their yang models. When both are loaded, we need to
specifically mention which connection-info we are referring
to in our rest body. We have done similar fix in the past
as well

https://git.opendaylight.org/gerrit/#/c/37216/

Change-Id: I1067b96619e0017b9ff7bb8c644fd915ba5836aa
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>