Update patch set 9
Patch Set 9:

I fixed what we talked about, I had to write new functions as - Create Networks, Delete Networks and some more because I needed to do iterated loop - on for the sites and one for the networks. So except that I did everything we talked about.

Patch-set: 9
Label: Verified=0