Update patch set 11
Patch Set 11:

> rather be well inserted in the appropriate test suite

+1.

> test case should not be named after a bug

That depends. If a scenario is non-trivial and described in the Bug, it is ok to use Bug number instead of otherwise long descriptive name.

In this case names line Two_Endpoints_Same_Name and Two_Endpoints_Different_Name might be both short and descriptive enough, so it is better to not use Bug in the names.

Patch-set: 11