Update patch set 11
Patch Set 11: Code-Review+1

> > rather be well inserted in the appropriate test suite
 > 
 > +1.
 > 
 > > test case should not be named after a bug
 > 
 > That depends. If a scenario is non-trivial and described in the
 > Bug, it is ok to use Bug number instead of otherwise long
 > descriptive name.
 > 
 > In this case names line Two_Endpoints_Same_Name and
 > Two_Endpoints_Different_Name might be both short and descriptive
 > enough, so it is better to not use Bug in the names.

Thanks for the review guys. This suite patch needs to get some legs and
get off my plate. it's been a review too long.

the entire purpose of this suite is for one-off bug automation to help
us know we don't get a regression in a bug that was found (probably in
some manual testing or reported from a 3rd party). Using the bug number
in the test name makes it clean and obvious when you are looking at
the test case level in the robot report. like this:

https://logs.opendaylight.org/sandbox/jenkins091/ovsdb-csit-1node-upstream-southbound-only-carbon/1/archives/log.html.gz#s1-s6

that job is sandbox proof of this patch as well as this one:

https://jenkins.opendaylight.org/sandbox/job/ovsdb-csit-1node-upstream-southbound-all-carbon/1/


btw, we are doing similar things in openflowplugin:

https://logs.opendaylight.org/releng/jenkins092/openflowplugin-csit-1node-flow-services-only-carbon/568/archives/log.html.gz#s1-s6

if it's a big deal to *not* write test automation this way, we'll have to
take this up as a larger refactor. I don't have any time for that at the
moment.

Patch-set: 11
Label: Code-Review=+1