Update patch set 14
Patch Set 14:

(3 comments)

this patch is looking nearly ready. maybe a few small updates.

I do think there is still a lot of duplicated code in the test
cases that we can still extract to a lib, but not for this patch.
let's get this one in, as it makes things much more clean and clear,
imho.

but I request that we get a bunch of sandbox runs with job(s) 
that sometimes see this dhcp bug. I think this one [a] should
work:

[a] netvirt-csit-1node-openstack-newton-nodl-v2-upstream-stateful-carbon

Patch-set: 14
Label: Verified=0