Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

This removes multiple features of the previous library, such as configurable username, password, prompt; log compression and downloading; and so on.

If you need multiple dumps at once, why not just do a small change for Start_* to return the new connection index, and Stop_* to accept it as an argument?

Patch-set: 3
Label: Code-Review=-1