Update patch set 1
Patch Set 1:

> Jamo, if you find even a single job that *doesn't* fail on this
 > test, we can give up skipping - but I think it consistently fails
 > now.
 > 
 > Revi we should probably add a link to the bug here, and also it
 > would be better if we could "skip_test" without any branch
 > condition.

we can skip because it does fail every time *now*, but I replied
somewhere (not here, but can't remember where) that we did see
this test fail sporadically before. So, likely there is a bug 
out there.

we need to get the environment figured out asap so we can re-enable
this test.

Patch-set: 1
Label: Code-Review=0