Update patch set 7
Patch Set 7:

(1 comment)

> Alon, I'm even more confused now.
 > 
 > here is the sandbox with the revert of this patch:
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-stateful-carbon/12/
 > 
 > the tests are run and they fail...
 > 
 > ???

What happened is this - you removed the ${tempest_config_file} parameter - causing public_network_id not to be filled in.

A side effect of this is a TON of tests are skipped (including our problematic two) - this has to be reverted soon.

After the revert, we're back to where we started - we still need to skip these two for mitaka.

Need to see what Mitaka ostestr has to offer in that regard - we can probably exclude these tests somehow - by the way, exclude only the two tests instead of the entire suite.

Patch-set: 7