Update patch set 46
Patch Set 46: Code-Review-1

(13 comments)

Some calls would read better if TemplatedRequests supported member index arguments. But it does not, so they are good enough for now.

I see MAC addresses are used, so this suites only work if ODL and TOOLS VMs are on the same LAN, right? If yes, document this restriction both in the suites and in the library.

Patch-set: 46
Label: Code-Review=-1