Update patch set 5
Patch Set 5:

> why do you reset the contribution date?

Personally, I like when dates in "git log" are not that far from merge dates. But also, the merged patch set is the last one, so the "contribution date" should refer to the last patch set, not the first one.

There are other views [0]. I just happen to be interested in the merge date more frequently than in the effort spent.

[0] https://lists.opendaylight.org/pipermail/integration-dev/2015-October/005059.html

Patch-set: 5