Update patch set 2
Patch Set 2: Code-Review-1

I do not understand this patch yet? it's very large to just be
concerned with moving back to getting them to pass with a boron
controller. I also see that the VERSION used is still "master"
which is currently carbon and soon to be nitrogen.

anyway, see my emails for some other discussion:
https://lists.opendaylight.org/pipermail/sfc-dev/2017-March/003972.html
https://lists.opendaylight.org/pipermail/sfc-dev/2017-March/003973.html

Patch-set: 2
Label: Code-Review=-1