Update patch set 3
Patch Set 3:

> (1 comment)
 > 
 > > Build Failed
 > >
 > > https://jenkins.opendaylight.org/releng/job/integration-test-verify-tox-master/655/
 > > : FAILURE
 > 
 > in this console log, you will find:
 > 
 > 16:44:41 Project wide:
 > 16:44:41 |    | [NORMAL] GitCommitBear:
 > 16:44:41 |    | Shortlog of the HEAD commit contains 51
 > character(s). This is 1 character(s) longer than the limit (51 >
 > 50).
 > 
 > just a slightly too long commit msg summary.
 > 
Thanks, I had totally oversight this
 > >
 > > Logs: https://logs.opendaylight.org/releng/jenkins092/integration-test-verify-tox-master/655
 > >
 > > https://jenkins.opendaylight.org/releng/job/integration-csit-verify-1node-library/2419/
 > > : UNSTABLE
 > >
 > > Logs: https://logs.opendaylight.org/releng/jenkins092/integration-csit-verify-1node-library/2419

Regarding executing the deletion two times: sudo execution is failing for me (local execution) but I was thinking that the Daexim guys added this sudo on purpose because they might need it on their scenario. The other option I though of was to use the "run with sudo, then without sudo if it fails" construct, but looked messier to me than this. If Daexim people confirm the sudo is not necessary, I will change it of course

Patch-set: 3
Label: Verified=0