Update patch set 1
Patch Set 1:

I have opened Bug 8138 as New Feature, as in a recent discussion Robert was telling me deleting snapshots and journal without renaming the member is incorrect.
In case developers decide 8138 is a real Bug and it should be fixed in Carbon, https://git.opendaylight.org/gerrit/54260 is there to decrease impact on other suites.

Patch-set: 1
Label: Verified=0