Update patch set 47
Patch Set 47:

> Mor, Jamo, want to start reviewing/merging this?
 > 2site jobs are busted anyways until this goes in - I think we can
 > review the bulk and add fixes in a new patch

I *quickly* glanced at the code and seems like it should be
reasonable to review/merge, but we need to first run a bunch of
the other openstack jobs that this will touch just to make sure
nothing is accidentally broken.

also, this needs a manual rebase.

Patch-set: 47
Label: Verified=0