Update patch set 9
Patch Set 9:

Hi jamo,

Any test is good.

But the gate should be a test that runs always faster and hence it should contain only blackbox tests.

Whitebox tests are better aligned only with the upstream job.  If we allow this checkin, it will start running in Gate increasing the time it takes for the Gate run to complete..

In general, my humble request is that we need to have two sets of suites:
1. Upstream suite that runs periodically and contains everything.
2. Gate suite that is run by developers per patch and does only black box testing.

Let me know what you think.

Vivek

Patch-set: 9