Update patch set 9
Patch Set 9:

> Hi jamo,
 > 
 > Any test is good.
 > 
 > But the gate should be a test that runs always faster and hence it
 > should contain only blackbox tests.
 > 
 > Whitebox tests are better aligned only with the upstream job.  If
 > we allow this checkin, it will start running in Gate increasing the
 > time it takes for the Gate run to complete..
 > 
 > In general, my humble request is that we need to have two sets of
 > suites:
 > 1. Upstream suite that runs periodically and contains everything.
 > 2. Gate suite that is run by developers per patch and does only
 > black box testing.
 > 
 > Let me know what you think.
 > 
 > Vivek

Hi Vivek,

I understand where you are coming from, but the trouble with our
situation is the length of these tests aren't really much of a 
factor here....yet. We currently spend aprox 35% of the entire
job time on setup (aprox 25m). So that is a consideration when
trying for a "short" gate for devs. When we get to many hours
worth of tests, I can see the argument for less runtime in gate
tests and more upstream.

I don't see much harm in having as much gating as possible right
down as close to the source as possible. In ODL, once a bad patch
gets merged and we slowly come to figure out something has broken,
it takes quite a bit of work to figure out which patch broke us.

I guess what I'm asking is how it is detrimental for a dev to
be blocked on a white box test failure? If it's because the 
white box tests rarely provide value and take a very long time
to come back, I could see the point, but that's not currently
the state.

maybe we have another DDF session around this?

JamO

Patch-set: 9