Backup-Restore support library + tests
The purpose of this library is to support the generic verification
of feature correctness in backup + restore scenarios.

A library is correct from a backup+restore standpoint when,
at any provisioning point, a controller backup, followed by
a restore, can be performed, and the execution of that
procedure will not have any impact on traffic tests or
datastore state check in respect to the expected behaviour
when no backup + restore procedures are performed.

For doing so, the library provides:

- A new keyword (BackupRestoreCheck), which:
  1) Performs a complete datastore export (using Daexim export rpc)
  2) Performs a backup, then a restore (not actual script here,
     since backup&restore mechanisms are deployment - dependant; a
     placeholder is provided for this step)
  3) Performs a new datastore export
  4) Compares both config & operational datastores for differences
     (that is, pre-backup and post-restore exports for both datastores),
     optionally prefiltering those exports using pre-filter files
  5) Fails when pre-backup and post-restore exports are different
     even after removing the specified filtered parts

- A new keyword (ConditionalBackupRestoreCheck), which performs
  the same steps than BackupRestoreCheck only when a command-line flag
  ("-v BR_TESTING_ENABLED:true") is present. This allows to easily
  add backup-restore verification on existing tests, allowing to
  toggle the execution of that verification

- Some unit tests for the library

jsonpatch library (https://pypi.python.org/pypi/jsonpatch) is expected to
be installed in order for this library to work

Note: for additional details regarding this lib, please check the
README.txt file included in the commit

Change-Id: Iad758ddc362cec319c7e843c1a6e4fcf728dad10
Signed-off-by: Diego Granados <diego.jesus.granados.lopez@ericsson.com>