Backup-Restore support library + tests
The purpose of this library is to allow the generic verification
of feature correctness in backup + restore scenarios.

A library is correct from a backup+restore standpoint when,
at any provisioning point, a controller backup, followed by
a restore, can be performed, and the execution of that
procedure will not have any impact on traffic tests or
datastore state check in respect to the expected behaviour
when no backup + restore procedures are performed.

For doing so, the library provides:

- A new keyword (BackupRestoreCheck), which:
  1) Performs a complete datastore export (using Daexim export rpc)
  2) Does a backup
  3) Restores the backup previously created
  4) Performs a new datastore export
  5) Compares both config & operational datastores for differences
     (that is, pre-backup and post-restore exports for both datastores),
     optionally prefiltering those exports using pre-filter files
  6) Fails when pre-backup and post-restore exports are different
     even after removing the specified filtered parts

- A new keyword (ConditionalBackupRestoreCheck), which performs
  the same steps than BackupRestoreCheck only when a command-line flag
  ("-v BR_TESTING_ENABLED:true") is present. This allows to easily
  add backup-restore verification on existing tests, allowing to
  toggle the execution of that verification

- Some tests for the library (based in SFC project)

The library includes in the commit itself the jsonpath library by Phil Budne
(https://pypi.python.org/pypi/jsonpath/). This had to be done in order to
rename the module (from jsonpath to jsonpathl), because ride fails to
import a class from a module with the same name, which is the case for
this library.

jsonpatch library (https://pypi.python.org/pypi/jsonpatch) is expected to
be installed in order for this library to work

Change-Id: Iad758ddc362cec319c7e843c1a6e4fcf728dad10
Signed-off-by: Diego Granados <diego.jesus.granados.lopez@ericsson.com>