Update patch set 76
Patch Set 76: Code-Review+1

So even though this contains most of work needed towards additional suites, it only does small changes to libraries. The idea is that the suites will be debugged in mutually independent changes, not requiring rebase over this.
Looks good, but I would like to see a Sandbox showing this does not break the currently run suites.

Patch-set: 76