Fix suites based on bug 8207
There is a time for akka to remove the node from the cluster. This time
is usually longer than the time when remaining/unisolated nodes elect
a new owner. The node which was and is follower still keeps a reference
to the isolated node until akka notification.
Wait is needed, WUSK were added to wait longer for cluster to work as
expected.

Change-Id: Ifda89ca6e4370282ef98c11ec5a837485c72279b
Signed-off-by: Peter Gubka <pgubka@cisco.com>