Add bgp suites with explicit shard leader location
Suites add 300k prefixes when rib owner is on the same/different
odl node as default/topology (config and operational) shard
leader.

In PrefixcountKeywords.robot the http sessions are created with
timeout=40. It is due to a big size of the http response when
all 300k or 1M prefixes are requested for counting.

Change-Id: I2e816df917402af71f51fe093553a072cc44e2a9
Signed-off-by: Peter Gubka <pgubka@cisco.com>