Give some life to AAA CSIT suite
I have no idea if/how/when this suite and it's library ever
worked. I am not trying to do anything here but make it work.
Locally, these changes give 13 pass, 4 failures. I think the
4 failures may be valid bugs, and will follow up with that.
Currently, upstream this suite is failing every test case and
has for as long as I know.

The python AAAJsonUtils.py library is not used by any other
suites so the changes will not affect anyone else.

Again, I was not looking to be pretty/nice/refactor/etc. I
just wanted to see this stuff start to work.

Change-Id: I3438e7e197d673b2afc37033cf446dfa57c5f9ff
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>