Update patch set 26
Patch Set 26: Code-Review+1

> (2 comments)
 > 
 > Sorry for the delay in estabilizing these test cases. The changes
 > from Jamo's comments are already implemented and I tried to make
 > the tests more robust as they were failing randomly.
 > 
 > Here you can see the sandbox results:
 > 
 > https://jenkins.opendaylight.org/sandbox/job/sfc-csit-3node-docker-full-deploy-all-carbon/

there is a consistent failure in the sandbox. Is there a known/open bug to account
for that? If not, and it's a valid problem, you can file a bug and use our robot
hooks (example: [0]) to denote that bug in the robot log. If the bug is getting fixed in the
short term, it's not worth the effort, imho. let me know and we can see about merging
this.

Thanks,
JamO

[0] https://github.com/opendaylight/integration-test/blob/master/csit/suites/ovsdb/Southbound_Domain/Bug_Validation.robot#L117

Patch-set: 26