Filter and reorder clustering suites
Prefix-based shards always use tell-based protocol,
so remove suites which test them with ask-based global configuration.

Ask-based protocol is known to produce AskTimeoutException
when isolating a member when transactions are in flight.
Singleton and Rpc tests are kept in for now,
will update when developers confirm
ask-based protocol is unreliable there too.

Otherwise make order correspond to the report document.

Change-Id: I47dd074fd50b3965a34c4568b1dbcc9baa292725
Signed-off-by: Vratko Polak <vrpolak@cisco.com>