Update patch set 1
Patch Set 1: Code-Review+1

this should get rid of the warnings we are now seeing, like 
here:

https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-newton-nodl-v2-upstream-stateful-carbon/544/archives/log.html.gz

I think it's actually a bug in robot framework to repor the 
warning. the warning is coming because Utils.robot now imports
the OperatingSystem library WITH NAME "os", and then complains
that "os" is a custom lib, and thinks it's different than 
OperatingSystem.

anyway, I just don't want to see the warnings any more.

Patch-set: 1
Label: Code-Review=+1