Update patch set 40
Patch Set 40:

> > Uploaded patch set 40.
 > 
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-jamo-upstream-stateful-carbon/1/

well, this sandbox job had a few failures. They may just be our 
sporadic failures, but I could not verify. The logging system is
broken ATM and they are not getting uploaded. I think it's my
fault for that and will get that fixed.

in the meantime we can try again, and if we get 100% pass we can
merge under the assumption that the job 1 above was just sporadic
failures.

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-jamo-upstream-stateful-carbon/2/

and an ocata try:

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-jamo-upstream-stateful-carbon/1/

Patch-set: 40