Enabled Debug log
Changed the log code

Output of Cache and waiting to remove all earlier tunnels from system
Moved the cache and waiting code at the beginning of first test case.

Changed the datastore dump.

Removed the debug logs of karaf as it is giving error
Sleep is put to identify the issue

Change-Id: I98328bd9582b6bc56377ec35cdb9027f424059e3
Signed-off-by: Nazeerahamed Hunshyal <nazeerahamed.hunshyal@tcs.com>