Update patch set 2
Patch Set 2:

Sridhar,

this is the patch I was referring to. I'm trying it in the sandbox
with the Carbon [0] and if it all passes we can merge right away.

if there are failures, we need to figure out what's wrong, but we
can merge a subset of these few v6 tempest cases.

JamO

[0] https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-jamo-upstream-stateful-carbon/2/

Patch-set: 2
Label: Verified=0