Update patch set 6
Patch Set 6:

Yes we should fix them but that can be done in separate patches. The problem is folks copy paste things from other scripts without always understanding what they are doing and then bad practice spreads around.

License headers MUST be at the top of the file (that's why it's called a license HEADER). The only exception is if the language has language specific parsing that must happen first for example shebang lines only work if they are the first line of the file. As well as the Python encoding string which needs to be 2nd if used.

Patch-set: 6
Label: Verified=0