Update patch set 28
Patch Set 28:

> > > the sandbox results are not as I was expecting (100% pass, or
 > > with
 > > > known sporadic bugs causing failures). I think two of the jobs
 > > > with failures didn't get their logs uploaded to nexus either
 > > (infra
 > > > problem going on right now)
 > >
 > > Failures in Ocata job are due to new cases ipv6 vpnservice cases.
 > > There seems to be command failures. The releng has not yet
 > executed
 > > these new cases.  I am retriggereing the jobs in sandbox again
 

I'd really love to merge this patch to help us save time, but we
need to make sure the results are going to be good. It's hard to
dig through existing failures as it is. I don't want to have more
showing up :)

 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-ocata-upstream-stateful-snat-conntrack-carbon/28/console

the failures here are expected for Ocata.
 
 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-upstream-learn-carbon/2/
 
there is one failure here that I don't know if it's right. it's in the test case
"Delete Router With NonExistentRouter Name" from the basic_vpnservice suite

the sandbox is going to
get purged soon, so pasting the little bit I can from the robot log here:

KEYWORD SSHLibrary . Write ${cmd};echo Command Returns $?
Documentation: 	

Writes the given `text` on the remote machine and appends a newline.
Start / End / Elapsed: 	20170622 22:13:13.912 / 20170622 22:13:13.918 / 00:00:00.006
22:13:13.918 	INFO 	neutron router-delete nonExistent 
Router;echo Command Returns $? 	
00:00:02.070 KEYWORD ${output} = SSHLibrary . Read Until Prompt
Documentation: 	

Consumes and returns the server output until the prompt is found.
Start / End / Elapsed: 	20170622 22:13:13.918 / 20170622 22:13:15.988 / 00:00:02.070
22:13:15.988 	INFO 	An auth plugin is required to fetch a token
Command Returns 1
[jenkins@sandbox-36636-2-devstack-newton-2 ~]> 	
22:13:15.988 	INFO 	${output} = An auth plugin is required to fetch a token
Command Returns 1
[jenkins@sandbox-36636-2-devstack-newton-2 ~]> 	
22:13:15.989 	INFO 	${output} = An auth plugin is required to fetch a token
Command Returns 1
[jenkins@sandbox-36636-2-devstack-newton-2 ~]> 	
00:00:00.115 KEYWORD SSHLibrary . Close Connection
Documentation: 	

Closes the current connection.
Start / End / Elapsed: 	20170622 22:13:15.989 / 20170622 22:13:16.104 / 00:00:00.115
00:00:00.002 KEYWORD BuiltIn . Should Match Regexp ${output}, Unable to find router with name or id 'nonExistentRouter'|Unable to find router\\(s\\) with id\\(s\\) 'nonExistentRouter'
Documentation: 	

Fails if string does not match pattern as a regular expression.
Start / End / Elapsed: 	20170622 22:13:16.104 / 20170622 22:13:16.106 / 00:00:00.002
22:13:16.105 	FAIL 	'An auth plugin is required to fetch a token
Command Returns 1
[jenkins@sandbox-36636-2-devstack-newton-2 ~]>' does not match 'Unable to find router with name or id 'nonExistentRouter'|Unable to find router\(s\) with id\(s\) 'nonExistentRouter''




The other failures in this job are the expected v6 failures because it's
in learn sg mode.


 > https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-newton-upstream-stateful-snat-conntrack-carbon/2/

actually, this job has the same failure in that router delete test case as
the above job.

I think once we get a handle on those, we can merge.

Patch-set: 28