Update patch set 3
Patch Set 3: Code-Review-1 Verified+1

> Hi Jamo,
 > 
 > Any update on results from this sandbox?
 > 
 > Vivek

There are still two failures, even with this patch, but I think
they are unrelated.

the error is that the arp reply group rule is not what it is 
expecting. I think the test wants to see the action to resubmit
to both 17 and 81, but it's now resubmitting to 48 and 81.

here's the error:

Keyword 'Verify GWMAC Flow Entry On Flow Table' failed after retrying for 30 seconds. The last error was: ' group_id=5000,type=all,bucket=actions=CONTROLLER:65535,bucket=actions=resubmit(,48),bucket=actions=resubmit(,81)' does not match 'actions=CONTROLLER:65535,bucket=actions=resubmit\(,17\),bucket=actions=resubmit\(,81\)'


so, with the new dhcp feature, have we changed where we are bouncing these
packets in the pipeline, no longer going to 17?

we can update that in this patch if so, and run in the sandbox. Otherwise,
if it's a bug we can merge this patch as is and deal with the bug. I scanned
through the feature description on github and I wasn't 100% sure.

Patch-set: 3
Label: Verified=+1
Label: Code-Review=-1