Update patch set 4
Patch Set 4:

I think it would be cleaner and easier to maintain to just selectively
skip the 2 test cases if Boron is the distro being tested.

you can add a line in the test case template keyword:

    Return From Keyword If    "skip_if_boron"" in @{TEST_TAGS}

then in the specific test cases, just add the tag "skip_if_boron"

that would save thousands of duplicated code in this patch.

check out these, for an example:

https://github.com/opendaylight/integration-test/blob/master/csit/suites/openstack/tempest/tempest.robot#L81

https://github.com/opendaylight/integration-test/blob/master/csit/libraries/DevstackUtils.robot#L23-L24

Patch-set: 4
Label: Code-Review=0