Change mask of OpenFlow flow 8 xml
As OVS treats ff:ff:ff:ff:ff:ff as no mask, when OpenFlowPlugin gets
statistics from OVS it do not receive any mask and then comparison
checks in CSIT tests are failing due to this, but it is correct
behaviour.
Change this mask to arbitrary mask so it will be deserialized to same
format instead of converted to prefix mask.

Change-Id: I7ec105047d1b80090b13caba4c93cc1df6af5d0d
Signed-off-by: Tomas Slusny <tomas.slusny@pantheon.tech>