Update patch set 2
Patch Set 2:

Hi Jamo, sure ;)

IMHO this goes into the right direction but I've some concernds. Basically what usually happens is that ostestr runs several threads at the same time so it can't debug. Adding the --pdb limits that to serial mode but you'd be in need of adding the actual pdb breakpoints to perform the debugging.

That said, we've recently merged this change [1] which I'd think you'd be interested into too. I'm not really sure about how do you perform the integration of tempest/odl but feel free to ping me on the irc (dmellado) for doubts.

Best!

[1] https://review.openstack.org/#/c/477471/

Patch-set: 2