Update patch set 9
Patch Set 9:

> How do we exit out of the pdb breakpoint, to continue to further
 > tempest scenario tests in CSIT (i.e,after we take the ODL resource
 > and flow dumps)?
 
one step at a time. Just getting it to pause before teardown is
my first step. I think I'm almost there, but I'm changing the actual
method of execution, so we might lose the ability to have an 
exclusion regex (that might be ok though)

anyway, when I get there, I'm hoping it's just a few keystrokes
at the pdb() prompt, like "continue" or something to let it finish.

if not, we can just CTRL-C out of it, and then I think we can 
manually call a tempest cleanup step, like Sridhar suggested.

 > I see this review uses a patch in tempest introduced recently:
 > http://git.openstack.org/cgit/openstack/tempest/commit/?id=215e82b894563f30e9aaa6c665f9b5db95e5666c

Right, if you scroll up in the comments you'll see that's what
Daniel pointed me to.

Patch-set: 9